requirements.txt
Flask==2.0.2
joblib==0.17.0
numpy==1.19.3
Pillow==8.4.0
tensorflow==2.7.0
gunicorn==20.0.4
Werkzeug==0.16.0
Jinja2==2.10.1
itsdangerous==1.1.0
MarkupSafe==1.1.1